requirements.txt
amqp==5.0.6
anyjson==0.3.3
argh==0.26.2
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
Django==3.1.7
django-capture-on-commit-callbacks==1.4.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.1
idna==2.10
importlib-metadata==3.10.1
iniconfig==1.1.1
kombu==5.0.2
Markdown==3.3.4
packaging==20.9
pathtools==0.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyobjc-core==7.2
pyobjc-framework-Cocoa==7.2
pyobjc-framework-FSEvents==7.2
pyparsing==2.4.7
pytest==6.2.3
pytest-celery==0.0.0a1
python-dateutil==2.8.1
pytz==2021.1
raises==0.2
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
test-plus==0.2.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
vine==5.0.0
watchdog==0.10.0
wcwidth==0.2.5
zipp==3.4.1
